Decouple summary route from docs renderer
